finmodel_app/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
pydantic==2.6.4
pydantic-settings==2.2.1
alembic==1.13.1
pandas==2.2.1
python-dateutil==2.9.0
jinja2==3.1.3
python-multipart==0.0.9